Updated bindgen and applied formatting suggestions from rustfmt and clippy.

+++ src/xxh3.rs
@@ -22,7 +22,7 @@ impl XXH3_64<'_> {
     /// One-shot hashing
     #[inline]
     pub fn hash(bytes: &[u8]) -> u64 {
-        unsafe { C::XXH3_64bits(bytes.as_ptr() as *const c_void, bytes.len() as u64) }
+        unsafe { C::XXH3_64bits(bytes.as_ptr() as *const c_void, bytes.len()) }
     }
 
     /// One-shot hashing with custom entropy buffer.
@@ -44,9 +44,9 @@ impl XXH3_64<'_> {
         assert!(entropy.len() >= (C::XXH3_SECRET_SIZE_MIN) as usize);
         C::XXH3_64bits_withSecret(
             bytes.as_ptr() as *const c_void,
-            bytes.len() as u64,
+            bytes.len(),
             entropy.as_ptr() as *const c_void,
-            entropy.len() as u64,
+            entropy.len(),
         )
     }
 
@@ -61,9 +61,7 @@ impl XXH3_64<'_> {
     /// One-shot hashing with seed
     #[inline]
     pub fn hash_with_seed(seed: u64, bytes: &[u8]) -> u64 {
-        unsafe {
-            C::XXH3_64bits_withSeed(bytes.as_ptr() as *const c_void, bytes.len() as u64, seed)
-        }
+        unsafe { C::XXH3_64bits_withSeed(bytes.as_ptr() as *const c_void, bytes.len(), seed) }
     }
 
     /// Streaming hashing
@@ -100,7 +98,7 @@ impl XXH3_64<'_> {
         C::XXH3_64bits_reset_withSecret(
             r.as_mut_ptr() as *mut C::XXH3_state_t,
             entropy.as_ptr() as *const c_void,
-            entropy.len() as u64,
+            entropy.len(),
         );
         XXH3_64 {
             state: r.assume_init(),
@@ -147,7 +145,7 @@ impl Hasher for XXH3_64<'_> {
             C::XXH3_64bits_update(
                 &mut self.state,
                 bytes.as_ptr() as *const c_void,
-                bytes.len() as u64,
+                bytes.len(),
             );
         }
     }
@@ -187,7 +185,7 @@ impl XXH3_128<'_> {
     /// One-shot hashing
     #[inline]
     pub fn hash(bytes: &[u8]) -> u128 {
-        let r = unsafe { C::XXH3_128bits(bytes.as_ptr() as *const c_void, bytes.len() as u64) };
+        let r = unsafe { C::XXH3_128bits(bytes.as_ptr() as *const c_void, bytes.len()) };
         xxh128_to_u128(r)
     }
 
@@ -210,9 +208,9 @@ impl XXH3_128<'_> {
         assert!(entropy.len() >= (C::XXH3_SECRET_SIZE_MIN) as usize);
         let r = C::XXH3_128bits_withSecret(
             bytes.as_ptr() as *const c_void,
-            bytes.len() as u64,
+            bytes.len(),
             entropy.as_ptr() as *const c_void,
-            entropy.len() as u64,
+            entropy.len(),
         );
         xxh128_to_u128(r)
     }
@@ -228,9 +226,8 @@ impl XXH3_128<'_> {
     /// One-shot hashing with seed
     #[inline]
     pub fn hash_with_seed(seed: u64, bytes: &[u8]) -> u128 {
-        let r = unsafe {
-            C::XXH3_128bits_withSeed(bytes.as_ptr() as *const c_void, bytes.len() as u64, seed)
-        };
+        let r =
+            unsafe { C::XXH3_128bits_withSeed(bytes.as_ptr() as *const c_void, bytes.len(), seed) };
         xxh128_to_u128(r)
     }
 
@@ -268,7 +265,7 @@ impl XXH3_128<'_> {
         C::XXH3_128bits_reset_withSecret(
             r.as_mut_ptr() as *mut C::XXH3_state_t,
             entropy.as_ptr() as *const c_void,
-            entropy.len() as u64,
+            entropy.len(),
         );
         XXH3_128 {
             state: r.assume_init(),
@@ -313,7 +310,7 @@ impl XXH3_128<'_> {
             C::XXH3_128bits_update(
                 &mut self.state,
                 bytes.as_ptr() as *const c_void,
-                bytes.len() as u64,
+                bytes.len(),
             );
         }
     }
